Correct compilation on createCoin

+++ apps/my-private-coin/klave/ERC20/ERC20Inputs.ts
@@ -4,7 +4,7 @@ import { JSON } from "@klave/sdk"
 export class CreateInput {
     name!: string;
     symbol!: string;
-    decimals!: u64;
+    decimals!: u8;
     totalSupply!: u64;
 }
